[ADD] 阻抗测量 undone

+++ Inc/protocol_ethernet.h
@@ -96,7 +96,7 @@ typedef struct
 void TCP_ProcessFSMInit(void);
 uint8_t TCP_ProcessFSM(void);
 
-uint8_t UDP_DataProcess(uint8_t SampleNum ,uint16_t Procesflag);
+uint8_t UDP_DataProcess(uint8_t SampleNum ,uint32_t Procesflag);
 uint8_t UDP_TriggerProcess();
 
 uint8_t protocol_RegisterAttrCBs(AttrCBs_t *pAttrcallbacks);
